+++ src/app/clients/page.jsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import Button from "@/component/Clients/ButtonClient";
+import Button from "@/component/Clients/AddClient";
 import Listclients from "@/component/Clients/List";
 
 const Client = () => {
